move gui pages into single file. For a project of this size, having 5 seperate files for each page is a bit much
