+++ run.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
 cd src
+
+# setup python virtual env
 if [ -d ".venv" ]; then
     source .venv/bin/activate
 else
@@ -8,5 +10,10 @@ else
     source .venv/bin/activate
 fi
 
+# create database if it doesnt exist
+if [ ! -f "automata.db" ]; then
+    python3 db_setup.py
+fi
+
 pip install graphviz bcrypt tk
-python3 page_control.py
+python3 gui.py
